requirements.txt
Flask
python-dotenv
google-generativeai
requests
PyMuPDF
beautifulsoup4
lxml
redis
rq
pandas
numpy
markdown
# gunicorn is for Linux deployments; on Windows use waitress or run via python
gunicorn
yfinance